requirements.txt
# requirements.txt - Diet-Bot MVP
# Tested with Python 3.11 on Google Cloud Run (April 2025)

fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
openai==1.23.0
firebase-admin==6.4.0
google-cloud-storage==2.15.1
httpx==0.27.0
pillow==10.3.0
langchain==0.1.16
pydantic==2.7.1
python-multipart==0.0.9
typer==0.12.3
rich==13.7.1
langchain-openai==0.1.3 